+++ asciimatics/effects.py
@@ -411,13 +411,13 @@ class _Star(object):
     Simple class to represent a single star for the Stars special effect.
     """
 
-    _star_chars = "..+..   ...x...  ...*...         "
-
-    def __init__(self, screen):
+    def __init__(self, screen, pattern):
         """
         :param screen: The Screen being used for the Scene.
+        :param pattern: The pattern to loop through
         """
         self._screen = screen
+        self._star_chars = pattern
         self._cycle = None
         self._old_char = None
         self._respawn()
@@ -464,19 +464,21 @@ class Stars(Effect):
     Add random stars to the screen and make them twinkle.
     """
 
-    def __init__(self, screen, count, **kwargs):
+    def __init__(self, screen, count, pattern="..+..   ...x...  ...*...         ", **kwargs):
         """
         :param screen: The Screen being used for the Scene.
         :param count: The number of starts to create.
+        :param pattern: The string pattern for the stars to loop through
 
         Also see the common keyword arguments in :py:obj:`.Effect`.
         """
         super(Stars, self).__init__(screen, **kwargs)
+        self._pattern = pattern
         self._max = count
         self._stars = []
 
     def reset(self):
-        self._stars = [_Star(self._screen) for _ in range(self._max)]
+        self._stars = [_Star(self._screen, self._pattern) for _ in range(self._max)]
 
     def _update(self, frame_no):
         for star in self._stars: